% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-app.R
\name{add_external_resources}
\alias{add_external_resources}
\title{Add external resources to UI head}
\usage{
add_external_resources()
}
\value{
HTML head tag
}
\description{
Adds external resources to shiny application from \code{propalloc}'s www
directory.
}
\section{Adds the following:}{
\enumerate{
\item \code{\link[shinyjs:useShinyjs]{shinyjs::useShinyjs()}}
\item \code{\link[shinyWidgets:useSweetAlert]{shinyWidgets::useSweetAlert()}}
\item \code{\link[rintrojs:introjsUI]{rintrojs::introjsUI()}}
\item Custom CSS - flipBox, DT styles, etc.
\item Custom JavaScript - SOV Module JS functionality, flipBox, etc.
}
}

\keyword{app}
