% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_allocate_expenses.R
\name{allocate_expenses}
\alias{allocate_expenses}
\title{Allocate expenses}
\usage{
allocate_expenses(dat, costs, weight_variable = c("tiv", "premium"))
}
\arguments{
\item{dat}{data frame with required columns: tiv and
allocated}

\item{costs}{renewal costs list}

\item{weight_variable}{String: one of \code{c("tiv", "premium")} indicating
which variable should be used to weight and allocate expenses}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Allocate expenses
}
