% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod_allocation_menu.R
\name{allocation_menu}
\alias{allocation_menu}
\title{Allocation Menu Module - Server}
\usage{
allocation_menu(input, output, session, inputs_menu_list)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{inputs_menu_list}{list of reactives derived in the \code{inputs_menu}
module.}
}
\value{
list of reactives corresponding to the allocation's derived datasets.
}
\description{
Allocation menu module's server function.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){

 }
}
}
\seealso{
\code{\link[shiny]{downloadHandler}},\code{\link[shiny]{reactive}},\code{\link[shiny]{callModule}},\code{\link[shiny]{observe}},\code{\link[shiny]{updateSliderInput}},\code{\link[shiny]{showNotification}},\code{\link[shiny]{reactiveValues}},\code{\link[shiny]{observeEvent}},\code{\link[shiny]{req}}
\code{\link[writexl]{write_xlsx}}
\code{\link[shinyjs]{visibilityFuncs}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate-joins}}
\code{\link[DT]{dataTableOutput}},\code{\link[DT]{datatable}},\code{\link[DT]{formatCurrency}},\code{\link[DT]{dataTableProxy}},\code{\link[DT]{editData}}
\code{\link[tidyselect]{starts_with}}
\code{\link[htmltools]{withTags}}
\code{\link[utils]{str}}
}
\keyword{module}
