% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-app_ui.R
\name{app_header_ui}
\alias{app_header_ui}
\title{App Header UI Function}
\usage{
app_header_ui()
}
\value{
HTML for a \link[shinydashboard:dashboardHeader]{dashboardHeader}
}
\description{
Adds header buttons and contacts.
}
