% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-app_ui.R
\name{app_sidebar_ui}
\alias{app_sidebar_ui}
\title{App Sidebar UI Function}
\usage{
app_sidebar_ui()
}
\value{
HTML for a \link[shinydashboard:dashboardSidebar]{dashboardSidebar}
}
\description{
App Sidebar UI Function
}
