% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labelling.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Labels}
\usage{
apply_labels(
  data,
  dict,
  from = "value",
  to = "value_label",
  by = "variable",
  names_from = "variable",
  names_to = "variable_label",
  dataset_name = NULL,
  ...
)
}
\arguments{
\item{data}{dataset to apply labeling on.}

\item{dict}{dictionary to use for application of labeling.}

\item{from}{a column name or position defining words or keys to be replaced}

\item{to}{a column name or position defining replacement values}

\item{by}{character or integer - which column in \code{dict} defines the
columns in \code{data} corresponding to each section of the \code{dict}.
Defaults to "variable".}

\item{names_from}{column name or position defining where to match names from.
Defaults to "variable".}

\item{names_to}{column name or position defining replacements for
\code{names(data)}. Defaults to "variable_label".}

\item{dataset_name}{(optional) name of dataset to filter \code{dict} by}

\item{...}{Passed to \code{\link[matchmaker]{match_df}}}
}
\value{
a labelled, display friendly \link[tibble:tibble-package]{tibble}
}
\description{
Apply Labels
}
\examples{
dictionary <- data.frame(
  var = c(rep("x1", 2),
          rep("x2", 3),
          rep("x3", 5)),
  var_lab = c(rep("Column 1", 2),
              rep("Column 2", 3),
              rep("Column 3", 5)),
  val = c(c(TRUE, FALSE),
          c(1:3),
          letters[1:5]),
  val_lab = c(c("YES", "NO"),
              paste0("Group ", c(1:3)),
              paste0("Area ", LETTERS[1:5]))
)

dat <- data.frame(
  "x1" = rep(c(TRUE, FALSE), 15),
  "x2" = rep(c(1:3), 10),
  "x3" = rep(letters[1:5], 6)
)

apply_labels(dat,
             dictionary,
             from = "val",
             to = "val_lab",
             by = "var",
             names_from = "var",
             names_to = "var_lab")
}
