% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_ingest_rels.R
\name{apply_relativity}
\alias{apply_relativity}
\title{Apply Relativity}
\usage{
apply_relativity(
  relativity_data,
  coverage = c("aop", "cat_eq", "cat_wind", "cat_flood", "terrorism"),
  sov_linker = NA,
  sov = sov
)
}
\arguments{
\item{relativity_data}{df}

\item{coverage}{cov}

\item{sov_linker}{link}

\item{sov}{df}
}
\value{
df
}
\description{
Apply Relativity
}
