% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_apply_rels.R
\name{apply_rels}
\alias{apply_rels}
\title{Apply Relativity Adjustment Factors}
\usage{
apply_rels(
  bu_rels = NULL,
  sprinkler_tier_rels = NULL,
  combustible_rels = NULL,
  sov = NULL
)
}
\arguments{
\item{bu_rels}{BU rels}

\item{sprinkler_tier_rels}{Sprinkler Tier rels}

\item{combustible_rels}{Combustible rels}

\item{sov}{SOV}
}
\value{
tibble
}
\description{
Apply Relativity Adjustment Factors
}
