% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_apply_surcharges.R
\name{apply_surcharges}
\alias{apply_surcharges}
\title{Apply Claim Count Surcharges}
\usage{
apply_surcharges(preliminary_allocation_data, count_buckets)
}
\arguments{
\item{preliminary_allocation_data}{prelim data}

\item{count_buckets}{count buckets table}
}
\value{
data.frame
}
\description{
Apply Claim Count Surcharges
}
