% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_apply_threshold.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply Threshold}
\usage{
apply_threshold(dat, total_pct_chg, threshold)
}
\arguments{
\item{dat}{data frame with required columns: uncapped_allocated and
prior_allocated}

\item{total_pct_chg}{total percent change since prior on the overall
rate (premium / TIV) excluding expenses.}

\item{threshold}{threshold to apply}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Apply Threshold
}
