% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_alloc.R
\name{arrange_by_location}
\alias{arrange_by_location}
\title{Arrange a data.frame by location}
\usage{
arrange_by_location(data)
}
\arguments{
\item{data}{data containing a column for "location".}
}
\value{
an arranged tibble.
}
\description{
Since entity_id is a character field, this function parsed the number and
arranges the data in increasing order by the location's number.
}
