% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{bu_rels}
\alias{bu_rels}
\title{Relativities to apply by business unit}
\format{
A data frame with 2 rows and 6 columns that contains
Relativity to apply for indicated business unit.
\describe{
\item{bu}{String: Unique identifier representing the business unit of the entity}
\item{aop_bu_relativity}{AOP factor}
\item{cat_eq_bu_relativity}{CAT-EQ factor}
\item{cat_wind_bu_relativity}{CAT-Wind factor}
\item{cat_flood_bu_relativity}{CAT-Flood factor}
\item{terrorism_bu_relativity}{Terrorism factor}
}
}
\usage{
bu_rels
}
\description{
Relativities to apply by business unit
}
\keyword{datasets}
