% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{combustible_rels}
\alias{combustible_rels}
\title{All Other Peril relativities to apply by level of combustibility}
\format{
A data frame with 2 rows and 2 columns:
\describe{
\item{aop_combustible}{String: Unique identifier reflecting a level of combustibility}
\item{aop_combustible_relativity}{Numeric: Relativity to apply for indicated level of combustibility}
}
}
\usage{
combustible_rels
}
\description{
All Other Peril relativities to apply by level of combustibility
}
\keyword{datasets}
