% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{count_buckets}
\alias{count_buckets}
\title{Specification of surcharges to be applied to entities for claims made in the current financial period}
\format{
A data frame with 8 rows and 6 columns:
\describe{
\item{bucket}{String: Unique identifier of the bucket}
\item{name}{String: Display-quality descriptor of the bucket}
\item{min, max}{Numeric: Bounds of the bucket; bucket bounds should be mutually
exclusive, start at zero, and encapsulate the largest claim size}
\item{percent_surcharge, dollar_surcharge}{Numeric: The percentage of the claim,
or the dollar amount to be added to the premiums already allocated to the entity}
}
}
\usage{
count_buckets
}
\description{
Eventually left_join()ed on to \code{loss_run}.
}
\keyword{datasets}
