% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_count_buckets_tab.R
\name{count_buckets_tab}
\alias{count_buckets_tab}
\title{Claim count surcharges tab module - Server}
\usage{
count_buckets_tab(
  input,
  output,
  session,
  initial_count_buckets,
  dictionary,
  loss_run
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{initial_count_buckets}{count_buckets}

\item{dictionary}{dictionary}
}
\value{
reactive conut buckets data
}
\description{
Claim count surcharges tab module - Server
}
