% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_count_buckets_tab.R
\name{count_buckets_tab_ui}
\alias{count_buckets_tab_ui}
\title{Claim count surcharges tab module - UI}
\usage{
count_buckets_tab_ui(id)
}
\arguments{
\item{id}{shiny namespace ID}
}
\value{
HTML
}
\description{
Claim count surcharges tab module - UI
}
