% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{A dictionary that maps raw data column names and values to prettified strings for UI display}
\format{
A data frame with 1727 rows and 6 columns:
\describe{
\item{dataset}{String: Identifier of the dataset. Function \code{apply_labels()} takes as arguments
the data.frame of the raw data to be prettified, as well as the \code{dataset} string (amongst other arguments)
to which it is assigned. Therefore, the \code{dataset} string need not match exactly the symbol of the
data.frame of the raw data, although this might be the most-consistent approach. For example, this would
be a valid call:
\code{apply_labels(data = bu_rels, dict = propalloc::dictionary, dataset_name = "business_unit_rels")}
assuming that the dataset value \code{"business_unit_rels"} appears in propalloc::dictionary. The most-consistent
approach would be to use \code{"bu_rels"} as the \code{dataset} identifier, as in:
\code{apply_labels(data = bu_rels, dict = propalloc::dictionary, dataset_name = "bu_rels")}
}
\item{variable}{String: Field name of the dataset which is to be prettified. eg. "bu".}
\item{variable_label}{String: The prettified label to replace \code{variable} with for UI display. eg. "Business Unit".}
\item{value}{String: If \code{NA}, then this field is not used. Else, the dictionary will also attempt to prettify
values within the \code{variable} field. String values in this field should complete all possible values that the
\code{variable} can take. eg. "bu_a", "bu_b".}
\item{value_label}{String: The prettified label to replace \code{value} with for UI display.
eg. "Business Unit A", "Business Unit B". If \code{NA}, then this field is not used. If \code{value} field is \code{NA},
then this field should be too.}
\item{value_order}{Numeric: An integer value used to order the prettified data.frame according to the appearance of values in
field \code{value}.}
}
}
\usage{
dictionary
}
\description{
Used in conjunction with \code{propalloc::apply_labels()} to apply the dictionary's mapping to a data set
}
\keyword{datasets}
