% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{downloadButtonEdit}
\alias{downloadButtonEdit}
\title{Download Button without opening up new window in browser when downloaded}
\usage{
downloadButtonEdit(
  outputId,
  label = "Download",
  class = NULL,
  icon = NULL,
  ...
)
}
\arguments{
\item{outputId}{output ID}

\item{label}{label}

\item{class}{class}

\item{icon}{icon}

\item{...}{dots}
}
\value{
download bttn
}
\description{
Download Button without opening up new window in browser when downloaded
}
