% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_entity_loss_summary.R
\name{entity_loss_summary}
\alias{entity_loss_summary}
\title{Derive Loss Data by Entity}
\usage{
entity_loss_summary(
  loss_run,
  count_buckets,
  experience_period = NA,
  min_date = NA,
  max_date = NA
)
}
\arguments{
\item{loss_run}{loss run}

\item{count_buckets}{count bucket table}

\item{experience_period}{defined experience period}

\item{min_date}{minimum date}

\item{max_date}{maximum data}
}
\value{
tibble
}
\description{
This function derives total incurred, total claim counts, and claim counts
by "count bucket" for use in the allocation model.
}
