% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_extract_costs.R
\name{extract_costs}
\alias{extract_costs}
\title{Extract Costs for Allocation}
\usage{
extract_costs(cost_table)
}
\arguments{
\item{cost_table}{initial costs data.frame to extract from (default equals initial_costs)}
}
\value{
list of costs
}
\description{
This function takes as input the initial cost table (by default uses the internal
dataset \code{initial_costs}) and returns a named list containing:
\itemize{
\item terrorism - Terrorism Cost
\item all_risk - All Risk Cost (CAT + AOP)
\item risk_transfer - Risk Transfer Cost (CAT + AOP + Terrorism)
\item expenses - Total expenses
\item total_w_expense - Total Including Expenses
}
}
