% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{extract_number}
\alias{extract_number}
\title{Extract numbers from a string}
\usage{
extract_number(string, return_as = c("character", "numeric"))
}
\arguments{
\item{string}{String to pull numbers from}

\item{return_as}{Character or Numeric}
}
\value{
String of numbers
}
\description{
Extract numbers from a string
}
