% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{format_round_dollar}
\alias{format_round_dollar}
\title{Format a numeric as rounded dollars}
\usage{
format_round_dollar(value)
}
\arguments{
\item{value}{numeric}
}
\value{
character
}
\description{
Format a numeric as rounded dollars
}
\examples{
format_round_dollar(1234.56) # Returns "$1,235"
}
