% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_col_classes}
\alias{get_col_classes}
\title{Get column classes from a data.frame}
\usage{
get_col_classes(data)
}
\arguments{
\item{data}{data.frame input}
}
\value{
if each column only has one associated class, a named character vector
is returned with the names equal to the column names and values equal to
the classes. If some columns have more than a single class, a named list
with column names as names and classes as values is returned.
}
\description{
Get column classes from a data.frame
}
\examples{
data <- data.frame(a = c(1:3), b = letters[1:3], c = c(TRUE, FALSE, TRUE))
get_col_classes(data)

}
