% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{get_deps}
\alias{get_deps}
\title{Get Package Dependencies from DESCRIPTION}
\usage{
get_deps(
  path = "DESCRIPTION",
  dput = FALSE,
  field = c("Depends", "Imports", "Suggests")
)
}
\arguments{
\item{path}{current projects root path; defaults to your current working
directory.}

\item{dput}{logical}

\item{field}{Depends, Imports, or Suggests}
}
\value{
character vector of package names
}
\description{
Get Package Dependencies from DESCRIPTION
}
\examples{
\dontrun{
get_deps()
}
}
\keyword{internal}
