% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header_buttons.R
\name{header_buttons}
\alias{header_buttons}
\title{Header Buttons Server Module}
\usage{
header_buttons(
  input,
  output,
  session,
  parent_session = NULL,
  include_tour = TRUE,
  tour_steps = NULL,
  include_refresh = TRUE,
  include_help = TRUE,
  include_disclaimer = TRUE,
  help_path = fs::path_package("propalloc", "reports/help.Rmd"),
  disclaimer_path = fs::path_package("propalloc", "reports/disclaimer.Rmd")
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{parent_session}{shiny session of the parent environment where this
module is called from. Only used if \code{include_tour} is TRUE.}

\item{include_tour}{logical - include a 'tour app' button?}

\item{tour_steps}{list of 'steps' for tour. Passed to \code{\link[rintrojs:introjs]{rintrojs::introjs()}}.}

\item{include_refresh}{logical - include a 'refresh' button?}

\item{include_help}{logical - include a 'help' button?}

\item{include_disclaimer}{logical - include a 'disclaimer' button?}

\item{help_path}{character - if \code{include_help} is TRUE;
specify the path to the 'help' R Markdown document to utilize when button
is pressed.}

\item{disclaimer_path}{character - if \code{include_disclaimer} is TRUE;
specify the path to the 'disclaimer' R Markdown document to utilize when
button is pressed.}
}
\value{
server
}
\description{
Header Buttons Server Module
}
