% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_ingest_rels.R
\name{ingest_relativities}
\alias{ingest_relativities}
\title{Ingest Relativity Adjustment Factors}
\usage{
ingest_relativities(rels_list, sov = NULL)
}
\arguments{
\item{rels_list}{A multi-layered, named list of lists. The first level of lists
names are in the format
\code{name = list(applies_to = "<coverage>", data = <data.frame>)} where the
name refers to the name used to indentify the relativity (i.e. sprinkler tier,
business unit, etc), the applies_to defines which coverage to applu the
relativity adjustments to, and data is a 2-column data.frame where the first column
specifies character labels and the second column specifies numeric factors.}

\item{sov}{SOV}
}
\value{
tibble
}
\description{
This function consumes the user-defined relativity adjustment factors and
applies them to the TIV by entity resulting in a \link[tibble:tibble]{tibble::tibble} with
columns for each coverages relativity adjusted TIV.
}
