% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-mod_inputs_menu.R
\name{inputs_menu}
\alias{inputs_menu}
\title{Inputs tab Server}
\usage{
inputs_menu(input, output, session)
}
\arguments{
\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}
}
\value{
a named list of all input tables
}
\description{
Inputs tab Server
}
