% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-mod_inputs_menu.R
\name{inputs_menu_ui}
\alias{inputs_menu_ui}
\title{Inputs Tab Module UI}
\usage{
inputs_menu_ui(id)
}
\arguments{
\item{id}{namespace ID}
}
\value{
\code{\link[shinydashboard:tabItem]{shinydashboard::tabItem()}}
}
\description{
Inputs Tab Module UI
}
