% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-mod_insights_menu.R
\name{insights_menu}
\alias{insights_menu}
\title{Insights tab Server}
\usage{
insights_menu(input, output, session, allocation_data)
}
\arguments{
\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}
}
\value{
a named list of all input tables
}
\description{
Insights tab Server
}
