% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-mod_insights_menu.R
\name{insights_menu_ui}
\alias{insights_menu_ui}
\title{Insights Tab Module UI}
\usage{
insights_menu_ui(id)
}
\arguments{
\item{id}{namespace ID}
}
\value{
\code{\link[shinydashboard:tabItem]{shinydashboard::tabItem()}}
}
\description{
Insights Tab Module UI
}
