% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{load_demo_data}
\alias{load_demo_data}
\title{Load Demo Data for Use in Property Allocation Shiny App and Examples}
\usage{
load_demo_data(...)
}
\arguments{
\item{...}{passed to \code{\link[utils:data]{utils::data()}}}
}
\description{
Load Demo Data for Use in Property Allocation Shiny App and Examples
}
\examples{
library(propalloc)

\dontrun{
# unpack to global environment
load_demo_data(overwrite = TRUE)
}
}
