% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-mod_location_summary.R
\name{location_summary_menu}
\alias{location_summary_menu}
\title{Location Summary Menu Module - Server}
\usage{
location_summary_menu(input, output, session, allocation_data)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{allocation_data}{The \code{allocation_data} list of reactive data.frames
returned from \code{\link[=allocation_menu]{allocation_menu()}} 's server module.}
}
\value{
driver summary reactive list
}
\description{
Location Summary Menu Module - Server
}
