% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-mod_location_summary.R
\name{location_summary_menu_ui}
\alias{location_summary_menu_ui}
\title{Location Summary Menu UI}
\usage{
location_summary_menu_ui(id, sov)
}
\arguments{
\item{id}{ns id}
}
\value{
HTML
}
\description{
Location Summary Menu UI
}
