% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{loss_run}
\alias{loss_run}
\title{Claims made by entities in the current financial period}
\format{
A data frame with 482 rows and 8 columns:
\describe{
\item{claim_number}{String: Unique identifier of the claim}
\item{location_dud}{String: Unique location identifier}
\item{entity_id}{String: ID referencing one of \code{sov$entity_id} to which the claim applies}
\item{date_of_loss}{Date: Date of the claim}
\item{total_incurred}{Numeric: Size of the claim in dollars}
\item{accident_description}{String: Accident description}
\item{accident_location}{String: Accident Location}
\item{year}{Numeric: Year of Loss}
}
}
\usage{
loss_run
}
\description{
See \code{count_buckets}.
}
\keyword{datasets}
