% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loss_run_tab.R
\name{loss_run_tab}
\alias{loss_run_tab}
\title{Loss run tab module - Server}
\usage{
loss_run_tab(input, output, session, initial_loss_run, dictionary)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{initial_loss_run}{loss_run}

\item{dictionary}{dictionary}
}
\value{
reactive loss data
}
\description{
Loss run tab module - Server
}
