% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loss_run_tab.R
\name{loss_run_tab_ui}
\alias{loss_run_tab_ui}
\title{Loss run tab module - UI}
\usage{
loss_run_tab_ui(id)
}
\arguments{
\item{id}{shiny namespace ID}
}
\value{
HTML
}
\description{
Loss run tab module - UI
}
