% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_merge_entity_data.R
\name{merge_entity_data}
\alias{merge_entity_data}
\title{Derive Entity Data}
\usage{
merge_entity_data(
  sov,
  relativity_adjusted_tivs,
  entity_loss_data,
  rates,
  priors
)
}
\arguments{
\item{sov}{sov}

\item{relativity_adjusted_tivs}{rel adj tivs}

\item{entity_loss_data}{inc, buckets, and cnts by ent}

\item{rates}{rate table}

\item{priors}{prior premium table}
}
\value{
entity_data database
}
\description{
Derive Entity Data
}
