% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{picker}
\alias{picker}
\title{Picker Input}
\usage{
picker(
  id,
  label,
  choices,
  selected = choices,
  multiple = TRUE,
  count_threshold = 3,
  choice_options = NULL,
  ...
)
}
\arguments{
\item{id}{id}

\item{label}{label}

\item{choices}{choices}

\item{selected}{selected}

\item{multiple}{multiple}

\item{count_threshold}{count threshold}

\item{choice_options}{choce opts}

\item{...}{passed to \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}}}
}
\value{
HTML
}
\description{
Picker Input
}
