% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_preliminary_allocation.R
\name{preliminary_allocation}
\alias{preliminary_allocation}
\title{Perform Preliminary Allocation}
\usage{
preliminary_allocation(entity_data, costs, bg_pct)
}
\arguments{
\item{entity_data}{entity data}

\item{costs}{renewal costs list}

\item{bg_pct}{Budget Guidance Percentage (eg. 0.05 representing 5\%)}
}
\value{
data.frame
}
\description{
Perform Preliminary Allocation
}
