% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_driver_summary.R
\name{prepare_driver_summary}
\alias{prepare_driver_summary}
\title{Prepare a driver summary using the final allocation}
\usage{
prepare_driver_summary(allocated_data, filter_vector)
}
\arguments{
\item{allocated_data}{the data that has passed through all allocation stages
(\code{apply_threshold()} is the final stage). This data has had filters applied to
fields bu:department}

\item{filter_vector}{Named character vector. Entries represent the filter (from Rshiny) to be applied to
allocated_data. Values of the vector either NA (meaning no filter) or character. Names of vector
refer to bu:department}
}
\value{
data.frame
}
\description{
Prepare a driver summary using the final allocation
}
