% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{priors}
\alias{priors}
\title{Prior TIVs, premiums and expenses assessed for each entity}
\format{
A data frame with 739 rows and 12 columns:
\describe{
\item{entity_id}{String: Unique identifier of the entity; Set may not match those of \code{sov}}
\item{prior_tiv}{Numeric: The Total Insured Value assessed per entity in the prior financial period}
\item{prior_expenses}{Numeric: Expenses attributable to the entity in the prior financial period}
\item{prior_aop_premium, prior_cat_eq_premium, prior_cat_wind_premium, prior_cat_flood_premium, prior_terrorism_premium}{Numeric: Premiums attributable to lines of coverage for each entity}
\item{prior_total_cat_premium}{Numeric: The sum of premiums attributable to catastrophic lines of coverage only}
\item{prior_all_risk_premium}{Numeric: The sum of premiums attributable to all lines of coverage excluding terrorism}
\item{prior_risk_transfer_premium}{Numeric: The sum of premiums attributable to all lines of coverage}
\item{prior_premium_incl_expenses}{Numeric: The sum of all premiums and expenses for the entity in the prior financial period}
}
}
\usage{
priors
}
\description{
Prior TIVs, premiums and expenses assessed for each entity
}
\keyword{datasets}
