% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_priors_tab.R
\name{priors_tab}
\alias{priors_tab}
\title{Prior premiums tab module - Server}
\usage{
priors_tab(input, output, session, initial_priors, dictionary)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{initial_priors}{priors}

\item{dictionary}{dictionary}
}
\value{
reactive prior premiums data
}
\description{
Prior premiums tab module - Server
}
