% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_priors_tab.R
\name{priors_tab_ui}
\alias{priors_tab_ui}
\title{Prior premiums tab module - UI}
\usage{
priors_tab_ui(id)
}
\arguments{
\item{id}{shiny namespace ID}
}
\value{
HTML
}
\description{
Prior premiums tab module - UI
}
