% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_ingest_rels.R
\name{produce_factor}
\alias{produce_factor}
\title{Produce Factor - Internal}
\usage{
produce_factor(
  rel_data,
  coverage = c("aop", "cat_eq", "cat_wind", "cat_flood", "terrorism")
)
}
\arguments{
\item{rel_data}{df}

\item{coverage}{cov}
}
\value{
tibble
}
\description{
get all aop factors and derive total aop factor.
}
\keyword{internal}
