% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{rates}
\alias{rates}
\title{Rates which apply to TIVs to produce premiums}
\format{
A data frame with 38 rows and 5 columns:
\describe{
\item{rate_type}{String: One of \code{c("aop", "cat_eq", "cat_flood", "cat_wind", "terror")} reflecting the type of coverage; See vignette for details}
\item{rate_id}{String: ID reflecting the extent of the coverage of field \code{rate_type}; See vignette for details}
\item{prior_rate, market_rate, model_rate}{Numeric: Prior, market, and model rates of premium coverage relative to TIV}
}
}
\usage{
rates
}
\description{
The default rates appearing in this data frame are based on historical actuarial assessments
}
\keyword{datasets}
