% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rates_tab.R
\name{rates_tab}
\alias{rates_tab}
\title{Rates Tab Module - Server}
\usage{
rates_tab(input, output, session, initial_rates, dictionary)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{initial_rates}{initial rates}

\item{dictionary}{dictionary}
}
\description{
Rates Tab Module - Server
}
