% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_rebalance_premiums.R
\name{rebalance_premiums}
\alias{rebalance_premiums}
\title{Rebalance premiums to a total value}
\usage{
rebalance_premiums(
  the_data,
  column_name_to_rebalance,
  total_to_rebalance_to,
  do_keep_column_name = TRUE
)
}
\arguments{
\item{the_data}{data.frame}

\item{column_name_to_rebalance}{character reflecting the name of the (numeric) column that is being rebalanced}

\item{total_to_rebalance_to}{numeric}

\item{do_keep_column_name}{boolean If FALSE, then automatically renames to "column_name_rebalanced"}
}
\value{
data.frame
}
\description{
Preserves the portion of the total held by the column that is being rebalanced.
}
\examples{
data.frame(premiums = 1:3) \%>\% rebalance_premiums("premiums", 12)
# Returns data.frame(premiums = c(2, 4, 6))
}
