% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rels_tab.R
\name{rels_tab}
\alias{rels_tab}
\title{Relativities tab module - Server}
\usage{
rels_tab(
  input,
  output,
  session,
  initial_rels,
  dictionary,
  output_id = c("bu_rels_table", "sprinkler_tier_rels_table", "combustible_rels_table")
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{initial_rels}{bu_rels, sprinkler_tier_rels, combustible_rels}

\item{dictionary}{dictionary}

\item{output_id}{output id corresponding to the table generated.}
}
\value{
reactive relativity data
}
\description{
Relativities tab module - Server
}
