% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{renewal_costs}
\alias{renewal_costs}
\title{Premium and expense costs in prior and current financial years to allocate}
\format{
A data frame with 5 rows and 4 columns:
\describe{
\item{cost_type}{String: Either 'premium' or 'expense' indicating the type of cost}
\item{description}{String: Description the premium or expense}
\item{prior}{Numeric: The prior year's premium or expense}
\item{current}{Numeric: The current year's premium or expense}
}
}
\usage{
renewal_costs
}
\description{
Sets out the high-level premium and expense costs in
both prior and current years. The sum of premiums and expenses in prior
and current columns will be allocated to individual entities
i.e. the premiums allocated by the model will sum to equal the
current column of this table.
}
\details{
The table contains at least one entry with a description matching "all_risk"
and another matching "terrorism" - both of which should related to a
"premium" cost type.
}
\keyword{datasets}
