% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_renewal_costs_tab.R
\name{renewal_costs_tab}
\alias{renewal_costs_tab}
\title{Inputs Tab Module Server}
\usage{
renewal_costs_tab(
  input,
  output,
  session,
  initial_costs = renewal_costs,
  dictionary
)
}
\arguments{
\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}

\item{initial_costs}{data.frame for initial table representing renewal costs.}

\item{dictionary}{internal}
}
\value{
List of input data.frames
}
\description{
Inputs Tab Module Server
}
