% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_renewal_costs_tab.R
\name{renewal_costs_tab_ui}
\alias{renewal_costs_tab_ui}
\title{Renewal Costs Module UI}
\usage{
renewal_costs_tab_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
shinydashboard::tabBox HTML
}
\description{
Renewal Costs Module UI
}
