% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labelling.R
\name{reverse_labels}
\alias{reverse_labels}
\title{Reverse Labels}
\usage{
reverse_labels(
  data,
  dict,
  from = "value",
  to = "value_label",
  by = "variable",
  names_from = "variable",
  names_to = "variable_label",
  dataset_name = NULL,
  ...
)
}
\arguments{
\item{data}{dataset to apply labeling on.}

\item{dict}{dictionary to use for application of labeling.}

\item{from}{a column name or position defining words or keys to be replaced}

\item{to}{a column name or position defining replacement values}

\item{by}{character or integer - which column in \code{dict} defines the
columns in \code{data} corresponding to each section of the \code{dict}.
Defaults to "variable".}

\item{names_from}{column name or position defining where to match names from.
Defaults to "variable".}

\item{names_to}{column name or position defining replacements for
\code{names(data)}. Defaults to "variable_label".}

\item{dataset_name}{(optional) name of dataset to filter \code{dict} by}

\item{...}{Passed to \code{\link[matchmaker]{match_df}}}
}
\value{
un-labeled, R-friendly \link[tibble:tibble-package]{tibble}
}
\description{
Reverses \code{link[apply_labels()]}.
}
