% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{sort_numeric}
\alias{sort_numeric}
\title{Sort Numeric}
\usage{
sort_numeric(vec)
}
\arguments{
\item{vec}{character vector to sort.}
}
\value{
a sorted character vector, sorted by its extracted integers
}
\description{
Combines extract_number with sort to sort by integer values within character
strings. Useful to sort in order Unit 1, Unit 2, Unit 3 opposed to
Unit 1, Unit 10, Unit 11 (with 1's leading the default sort method).
}
\details{
If any strings do not contain a number they will be placed first by default.
}
\examples{
library(propalloc)

# undesired sort method
pull_unique(sov, "department") \%>\% sort_numeric()

# correct method
pull_unique(sov, "department") \%>\% sort_numeric()
}
