% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{sov}
\alias{sov}
\title{Current entity data detailing individual TIVs}
\format{
A data frame with 739 rows and 24 columns:
\describe{
\item{entity_id}{String: Unique identifier of the entity}
\item{loss_run_id}{String: Identifier to match losses to entities.}
\item{bu, region, country, state, division, location, department}{String: Attributes describing the entity. bu = Business unit.}
\item{tiv}{Numeric: The Total Insured Value assessed per entity}
\item{aop_coverage, cat_eq_coverage, cat_wind_coverage, cat_flood_coverage, terrorism_coverage}{Boolean: Indicating whether entity takes such coverage}
\item{aop_sprinkler_tier}{String: Referencing exactly one of the tiers present in aop_sprinkler_tier, reflecting the tier of sprinkler system installed by the entity.}
\item{aop_combustible}{String: Referencing exactly one of the tiers present in combustible_rels, reflecting the level of combustibility of the entity.}
\item{aop_tiv_size_bucket}{String: Derivative field reflecting the relative size of TIV. Not used in default setup of the model, although can be conveniently used by custom rates (see \code{rates})}
\item{cat_wind_hurricane}{String: Referencing...}
\item{aop_id}{String: A geographical ID used by \code{rates}}
\item{cat_eq_id, cat_wind_id, cat_flood_id}{String: An ID representing the threat of earthquake-, wind-, or flood-related catastrophic damage used by \code{rates}}
\item{terrorism_id}{String: A geographical ID used by \code{rates}}
}
}
\usage{
sov
}
\description{
Sets out the high-level premium and expense costs in
both prior and current years. The sum of premiums and expenses in prior
and current columns will be allocated to individual entities
ie. the premiums allocated by the model will sum to equal the
current column of this table.
}
\details{
The table contains at least one entry with a description matching "all_risk"
and another matching "terrorism" - both of which should related to a
"premium" cost type.
}
\keyword{datasets}
