% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sov_tab.R
\name{sov_tab}
\alias{sov_tab}
\title{SOV Tab Module - Server}
\usage{
sov_tab(input, output, session, initial_sov, dictionary)
}
\arguments{
\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}

\item{initial_sov}{data.frame for initial table representing sov.}

\item{dictionary}{internal}
}
\value{
sov table
}
\description{
SOV Tab Module - Server
}
