% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sov_tab.R
\name{sov_tab_ui}
\alias{sov_tab_ui}
\title{SOV Tab Module - UI}
\usage{
sov_tab_ui(id)
}
\arguments{
\item{id}{Shiny Namespace ID}
}
\value{
HTML for a \link[shiny:tabPanel]{shiny::tabPanel} (for the \link[shinydashboard:tabBox]{shinydashboard::tabBox})
}
\description{
This function defines the user-interface behind the Schedule of Values (SOV)
input tab within the shiny application.
}
