% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{sprinkler_tier_rels}
\alias{sprinkler_tier_rels}
\title{All Other Peril relativities to apply by tier of sprinkler system}
\format{
A data frame with 4 rows and 2 columns:
\describe{
\item{aop_sprinkler_tier}{String: Unique identifier reflecting a tier of sprinkler system}
\item{aop_sprinkler_tier_relativity}{Numeric: Relativity to apply for indicated tier of sprinkler system}
}
}
\usage{
sprinkler_tier_rels
}
\description{
All Other Peril relativities to apply by tier of sprinkler system
}
\keyword{datasets}
