% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_package.R
\docType{package}
\name{propalloc-package}
\alias{propalloc}
\alias{propalloc-package}
\title{propalloc: Oliver Wyman RShiny Property Allocation App}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Property Allocation model in R Shiny that allocates
    an insured's prospective year property insurance costs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/propalloc}
  \item Report bugs at \url{https://github.com/jimbrig/propalloc/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

Other contributors:
\itemize{
  \item Jack Reardon \email{jack.reardon@oliverwyman.com} [contributor]
  \item Adam Lewis \email{adam.lewis@oliverwyman.com} [reviewer]
  \item Oliver Wyman Actuarial Consulting [funder]
}

}
\keyword{internal}
