% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(install_dependencies = FALSE)
}
\arguments{
\item{install_dependencies}{logical-install package/app dependencies?
Defaults to FALSE.}
}
\description{
Run the Shiny Application
}
\keyword{app}
